requirements.txt
coverage
Flask
pytest
gunicorn
mysql-connector-python
